Created an end game system that lets you view the game results

+++ src/main/java/lycanthrope/repositories/PlayerRepository.java
@@ -3,5 +3,5 @@ package lycanthrope.repositories;
 import lycanthrope.models.Player;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlayerRepository extends CrudRepository<Player, Long> {
+public interface PlayerRepository extends CrudRepository<Player, Integer> {
 }
